JIRA-135 #refac #comment "vector value 크기가 64bit이므로 float->double 변경"

+++ backend/epimetheus/src/main/java/uos/capstone/epimetheus/adapter/LlamaAdapter.java
@@ -3,10 +3,9 @@ package uos.capstone.epimetheus.adapter;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import uos.capstone.epimetheus.dtos.LlamaStepResponse;
-import uos.capstone.epimetheus.dtos.LlamaVectorResponse;
 
 public interface LlamaAdapter {
     Flux<LlamaStepResponse> getAllTaskSteps(String json);
 
-    Mono<float[]> getVectorFromSentence(String sentence);
+    Mono<double[]> getVectorFromSentence(String sentence);
 }
